Refactor opengl usage for apps and utit tests. All opengl code used by command line apps (and gpu tests) has been refactor. Ociodisplay is also using the refactored code and still uses  opengl directly. - add gpu support to ociochecklut. - add optimization level to ociodisplay.
