+++ src/apps/ociochecklut/CMakeLists.txt
@@ -1,6 +1,13 @@
 # SPDX-License-Identifier: BSD-3-Clause
 # Copyright Contributors to the OpenColorIO Project.
 
+if(NOT OCIO_GL_ENABLED)
+    message(WARNING "GL component missing. GPU disabled for ociochecklut.")
+else()
+    add_compile_definitions(OCIO_GPU_ENABLED)
+    set(OCIO_GL_LIB oglapphelpers)
+endif()
+
 set(SOURCES
     main.cpp
 )
@@ -23,6 +30,7 @@ set_target_properties(ociochecklut PROPERTIES
 
 target_link_libraries(ociochecklut 
     PRIVATE 
+        ${OCIO_GL_LIB}
         apputils
         OpenColorIO
 )
