+++ src/libutils/CMakeLists.txt
@@ -5,9 +5,10 @@ add_subdirectory(apphelpers)
 
 if(TARGET test_gpu_exec OR
    TARGET ociodisplay OR
+   TARGET ociochecklut OR
    TARGET ocioconvert
 )
-    add_subdirectory(oglbuilder)
+    add_subdirectory(oglapphelpers)
 endif()
 
 if(TARGET ocioconvert OR
