+++ tests/gpu/CMakeLists.txt
@@ -1,24 +1,9 @@
 # SPDX-License-Identifier: BSD-3-Clause
 # Copyright Contributors to the OpenColorIO Project.
 
-find_package(OpenGL QUIET)
-if(NOT OpenGL_FOUND)
-	message(WARNING "OpenGL not found. Skipping the GPU unit tests.")
-	return()
-endif()
-
-if(NOT APPLE)
-	find_package(GLEW QUIET)
-	if(NOT GLEW_FOUND)
-		message(WARNING "GLEW not found. Skipping the GPU unit tests.")
-		return()
-	endif()
-endif()
-
-find_package(GLUT QUIET)
-if(NOT GLUT_FOUND)
-	message(WARNING "GLUT not found. Skipping the GPU unit tests.")
-	return()
+if(NOT OCIO_GL_ENABLED)
+    message(WARNING "GL component missing. Skipping the GPU unit tests.")
+    return()
 endif()
 
 set(SOURCES
@@ -37,12 +22,6 @@ set(SOURCES
 )
 
 add_executable(test_gpu_exec ${SOURCES})
-target_include_directories(test_gpu_exec
-	PRIVATE
-		${OPENGL_INCLUDE_DIR}
-		${GLEW_INCLUDE_DIRS}
-		${GLUT_INCLUDE_DIR}
-)
 
 if(NOT BUILD_SHARED_LIBS)
 	target_compile_definitions(test_gpu_exec
@@ -69,11 +48,8 @@ set_target_properties(test_gpu_exec PROPERTIES
 target_link_libraries(test_gpu_exec
 	PRIVATE
 		OpenColorIO
-		oglbuilder
+		oglapphelpers
 		unittest_data
-		${OPENGL_LIBRAIRIES}
-		${GLEW_LIBRARIES}
-		${GLUT_LIBRARIES}
 )
 
 add_test(test_gpu test_gpu_exec)
