+++ tests/gpu/Lut3DOp_test.cpp
@@ -13,9 +13,6 @@
 namespace OCIO = OCIO_NAMESPACE;
 
 
-const int LUT3D_EDGE_SIZE = 32;
-
-
 OCIO_ADD_GPU_TEST(Lut3DOp, red_only_using_CSP_file_legacy_shader)
 {
     // Any other 3D LUT file format would have been good also.
@@ -53,10 +50,9 @@ OCIO_ADD_GPU_TEST(Lut3DOp, red_only_using_CSP_file_legacy_shader)
     file->setSrc(filename.c_str());
     file->setInterpolation(OCIO::INTERP_LINEAR);
 
-    OCIO::GpuShaderDescRcPtr shaderDesc
-        = OCIO::GpuShaderDesc::CreateLegacyShaderDesc(LUT3D_EDGE_SIZE);
-    test.setContext(file->createEditableCopy(), shaderDesc);
+    test.setProcessor(file);
 
+    test.setLegacyShader(true);
     test.setErrorThreshold(2e-4f);
 }
 
@@ -97,10 +93,9 @@ OCIO_ADD_GPU_TEST(Lut3DOp, green_only_using_CSP_file_legacy_shader)
     file->setSrc(filename.c_str());
     file->setInterpolation(OCIO::INTERP_LINEAR);
 
-    OCIO::GpuShaderDescRcPtr shaderDesc
-        = OCIO::GpuShaderDesc::CreateLegacyShaderDesc(LUT3D_EDGE_SIZE);
-    test.setContext(file->createEditableCopy(), shaderDesc);
+    test.setProcessor(file);
 
+    test.setLegacyShader(true);
     test.setErrorThreshold(2e-4f);
 }
 
@@ -141,10 +136,9 @@ OCIO_ADD_GPU_TEST(Lut3DOp, blue_only_using_CSP_file_legacy_shader)
     file->setSrc(filename.c_str());
     file->setInterpolation(OCIO::INTERP_LINEAR);
 
-    OCIO::GpuShaderDescRcPtr shaderDesc
-        = OCIO::GpuShaderDesc::CreateLegacyShaderDesc(LUT3D_EDGE_SIZE);
-    test.setContext(file->createEditableCopy(), shaderDesc);
+    test.setProcessor(file);
 
+    test.setLegacyShader(true);
     test.setErrorThreshold(2e-4f);
 }
 
@@ -186,10 +180,9 @@ OCIO_ADD_GPU_TEST(Lut3DOp, arbitrary_using_CSP_file_legacy_shader)
     file->setSrc(filename.c_str());
     file->setInterpolation(OCIO::INTERP_LINEAR);
 
-    OCIO::GpuShaderDescRcPtr shaderDesc
-        = OCIO::GpuShaderDesc::CreateLegacyShaderDesc(LUT3D_EDGE_SIZE);
+    test.setProcessor(file);
 
-    test.setContext(file->createEditableCopy(), shaderDesc);
+    test.setLegacyShader(true);
     test.setErrorThreshold(2e-4f);
 }
 
@@ -229,8 +222,7 @@ OCIO_ADD_GPU_TEST(Lut3DOp, arbitrary_using_CSP_file)
     file->setSrc(filename.c_str());
     file->setInterpolation(OCIO::INTERP_LINEAR);
 
-    OCIO::GpuShaderDescRcPtr shaderDesc  = OCIO::GpuShaderDesc::CreateShaderDesc();
-    test.setContext(file->createEditableCopy(), shaderDesc);
+    test.setProcessor(file);
 
     // TODO: Small LUTs not being resampled for now, such error threshold is expected
     //       The legacy shader has a better error threshold because
@@ -274,10 +266,9 @@ OCIO_ADD_GPU_TEST(Lut3DOp, 3dlut_file_legacy_shader)
 {
     OCIO::FileTransformRcPtr file = GetFileTransform("lut3d_1.spi3d");
 
-    OCIO::GpuShaderDescRcPtr shaderDesc
-        = OCIO::GpuShaderDesc::CreateLegacyShaderDesc(LUT3D_EDGE_SIZE);
-    test.setContext(file->createEditableCopy(), shaderDesc);
+    test.setProcessor(file);
 
+    test.setLegacyShader(true);
     test.setErrorThreshold(5e-4f);
 }
 
@@ -286,8 +277,7 @@ OCIO_ADD_GPU_TEST(Lut3DOp, 3dlut_file_spi3d_linear)
     // Linear interpolation
     OCIO::FileTransformRcPtr file = GetFileTransform("lut3d_1.spi3d");
 
-    OCIO::GpuShaderDescRcPtr shaderDesc = OCIO::GpuShaderDesc::CreateShaderDesc();
-    test.setContext(file->createEditableCopy(), shaderDesc);
+    test.setProcessor(file);
 
     test.setErrorThreshold(5e-4f);
 }
@@ -297,8 +287,7 @@ OCIO_ADD_GPU_TEST(Lut3DOp, 3dlut_file_spi3d_tetra)
     OCIO::FileTransformRcPtr file = GetFileTransform("lut3d_1.spi3d");
     file->setInterpolation(OCIO::INTERP_TETRAHEDRAL);
 
-    OCIO::GpuShaderDescRcPtr shaderDesc = OCIO::GpuShaderDesc::CreateShaderDesc();
-    test.setContext(file->createEditableCopy(), shaderDesc);
+    test.setProcessor(file);
 
     test.setErrorThreshold(1e-6f);
 
@@ -320,8 +309,7 @@ OCIO_ADD_GPU_TEST(Lut3DOp, inv3dlut_file_spi3d_linear)
     OCIO::FileTransformRcPtr file = GetFileTransform("lut3d_1.spi3d");
     file->setDirection(OCIO::TRANSFORM_DIR_INVERSE);
 
-    OCIO::GpuShaderDescRcPtr shaderDesc = OCIO::GpuShaderDesc::CreateShaderDesc();
-    test.setContext(file->createEditableCopy(), shaderDesc);
+    test.setProcessor(file);
 
     test.setErrorThreshold(1.2e-3f);
 }
@@ -337,8 +325,7 @@ OCIO_ADD_GPU_TEST(Lut3DOp, inv3dlut_file_spi3d_tetra)
     // inverse LUT, so this test produces the same result as the previous one.
     file->setInterpolation(OCIO::INTERP_TETRAHEDRAL);
 
-    OCIO::GpuShaderDescRcPtr shaderDesc = OCIO::GpuShaderDesc::CreateShaderDesc();
-    test.setContext(file->createEditableCopy(), shaderDesc);
+    test.setProcessor(file);
 
     test.setErrorThreshold(1.2e-3f);
 }
@@ -348,8 +335,7 @@ OCIO_ADD_GPU_TEST(Lut3DOp, 3dlut_file_spi3d_bizarre_linear)
     // Linear interpolation
     OCIO::FileTransformRcPtr file = GetFileTransform("lut3d_bizarre.spi3d");
 
-    OCIO::GpuShaderDescRcPtr shaderDesc = OCIO::GpuShaderDesc::CreateShaderDesc();
-    test.setContext(file->createEditableCopy(), shaderDesc);
+    test.setProcessor(file);
 
     // This is due to the fact that the LUT is small and to
     // the GPU 8-bit index quantization.
@@ -361,8 +347,7 @@ OCIO_ADD_GPU_TEST(Lut3DOp, 3dlut_file_spi3d_bizarre_tetra)
     OCIO::FileTransformRcPtr file = GetFileTransform("lut3d_bizarre.spi3d");
     file->setInterpolation(OCIO::INTERP_TETRAHEDRAL);
 
-    OCIO::GpuShaderDescRcPtr shaderDesc = OCIO::GpuShaderDesc::CreateShaderDesc();
-    test.setContext(file->createEditableCopy(), shaderDesc);
+    test.setProcessor(file);
 
     test.setErrorThreshold(1e-6f);
 
@@ -378,8 +363,7 @@ OCIO_ADD_GPU_TEST(Lut3DOp, inv3dlut_file_spi3d_bizarre_linear)
     OCIO::FileTransformRcPtr file = GetFileTransform("lut3d_bizarre.spi3d");
     file->setDirection(OCIO::TRANSFORM_DIR_INVERSE);
 
-    OCIO::GpuShaderDescRcPtr shaderDesc = OCIO::GpuShaderDesc::CreateShaderDesc();
-    test.setContext(file->createEditableCopy(), shaderDesc);
+    test.setProcessor(file);
 
     test.setErrorThreshold(3e-4f);
 }
@@ -393,8 +377,7 @@ OCIO_ADD_GPU_TEST(Lut3DOp, inv3dlut_file_spi3d_bizarre_tetra)
     file->setDirection(OCIO::TRANSFORM_DIR_INVERSE);
     file->setInterpolation(OCIO::INTERP_TETRAHEDRAL);
 
-    OCIO::GpuShaderDescRcPtr shaderDesc = OCIO::GpuShaderDesc::CreateShaderDesc();
-    test.setContext(file->createEditableCopy(), shaderDesc);
+    test.setProcessor(file);
 
     test.setErrorThreshold(3e-4f);
 }
